requirements.txt
deta==1.1.0
python-dotenv==0.21.1
slack-sdk==3.19.5